Move downloadSample logic to DownloadingProvider
